# swagger/swagger_docs.yaml
openapi: 3.1.0
info:
  title: Authorization API
  description: This is the API documentation for Authorization API
  version: 1.0.0
servers:
  - url: http://localhost:8002
    description: Main backend server
paths:
  /auth/sign-up:
    post:
      summary: Registrating and creating User
      operationId: SignUp
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: Unique identifier of User
                    example: 1
        '500':
          description: Server error
  /auth/sign-in:
    post:
      summary: Authorization for users
      operationId: SignIn
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: Email of user
                  example: omarzhakman@gmail.com
                password:
                  type: string
                  description: User's password
                  example: root
      responses:
        '200':
          description: Authentification is succefull
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token for User
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3MTc4NDA1NTMsImlhdCI6MTcxNzc5NzM1Mywicm9sZSI6ImFkbWluIiwidXNlcl9pZCI6NH0.Y3vgMKoYMLrlFacRSHlCw6kzm5nU62EkVmnT8PI6CNc
        '500':
          description: Server error


components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required:
        - id
        - email
        - phone
        - password
        - role
      properties:
        id:
          type: integer
          description: Unique identifier of User
          example: 1
        email:
          type: string
          description: Email of the user
          example: omarzhakman@gmail.com
        phone:
          type: string
          description: Phone number of the user
          example: +7 701 143 3567
        password:
          type: string
          description: Password of the User for Authorization
          example: root
        role:
          type: string
          description: Role of the user
          example: admin

    